Proposal_gen_Backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
jinja2
python-multipart
